SHPIA-API/nordic-api-master/requirements.txt
Django==2.2
djangorestframework==3.12.2
PyJWT==1.7.1
pymongo==3.11.4
pytz==2020.4
sqlparse==0.2.4
djongo==1.3.6
